migrate test_bluez_bluetooth to pytest
